==> 01_Preprocesamiento_imputacion_datos_geograficos_demograficos.ipynb <==
import pandas as pd
import numpy as np
from scipy.spatial import cKDTree
from prettytable import PrettyTable

bdif = pd.read_excel('bdif_pre.xlsx')
bdif.head()

# Copiamos las coordenadas de los municipios
bdif['lon'] = bdif['lon_ETRS89']
bdif['lat'] = bdif['lat_ETRS89']

# Necesitamos cargar el dataset del NGMEP
municipios = pd.read_excel('municipiosNGMEP.xlsx')
municipios.head()

# Para los registros que no tengan coordenadas buscaremos los mas cercanos
# Guardamos las coordenadas de los municipios y construimos el árbol
coordenadas_municipios = municipios.dropna(subset=['LONGITUD_ETRS89', 'LATITUD_ETRS89'])
tree = cKDTree(coordenadas_municipios[['LONGITUD_ETRS89', 'LATITUD_ETRS89']])

# Extraemos los registros que no tienen coordenadas
bdif_faltantes = bdif[bdif['lon'].isna() & bdif['lat'].isna()]

# Para las filas sin lon y lat, buscar el más cercano
for indice, fila in bdif_faltantes.iterrows():
    if not pd.isna(fila['x_lon']) and not pd.isna(fila['y_lat']):
        
        # Buscamos las coordenadas más cercanas
        distancia, ind = tree.query([fila['x_lon'], fila['y_lat']])

        # Tomamos los datos del municipio más cercano
        cercano = coordenadas_municipios.iloc[ind]

        # Imputa los valores en bdif
        bdif.at[indice, 'lon'] = cercano['LONGITUD_ETRS89']
        bdif.at[indice, 'lat'] = cercano['LATITUD_ETRS89']
        bdif.at[indice, 'cod_ine'] = cercano['COD_INE']
        bdif.at[indice, 'cod_prov'] = cercano['COD_PROV']
        bdif.at[indice, 'id_rel'] = cercano['ID_REL']
        bdif.at[indice, 'provincia'] = cercano['PROVINCIA']
        bdif.at[indice, 'nombre'] = cercano['NOMBRE_ACTUAL']
        bdif.at[indice, 'superficie'] = cercano['SUPERFICIE']
        bdif.at[indice, 'altitud'] = cercano['ALTITUD']
        bdif.at[indice, 'poblacion'] = cercano['POBLACION_MUNI']

bdif.columns

columnas = ['parte','año','cod_com','cod_prov','probignicion','diastormenta','diasultimalluvia','tempmaxima',
            'humrelativa','velocidadviento','direccionviento','detectadopor','deteccion','combustible','tipofuego',
            'lugar','mes','tipodia','id_rel','provincia','poblacion','superficie','altitud','areaquemada','lon','lat']
bdif = bdif[columnas]
# Exportar a un nuevo archivo
bdif[columnas].to_excel('bdif_geograficas.xlsx', index=False)

### Tabla de resultados

# Obtener dimensiones de la base de datos
filas, columnas = bdif.shape

# Creamos la tabla
tabla = PrettyTable(['Variable', 'No nulos', 'Tipo'])

# Añadimos los datos
for columna in bdif.columns:
    no_nulos = bdif[columna].notnull().sum()
    tipo = bdif[columna].dtype
    tabla.add_row([columna, no_nulos, tipo])

# Info de las dimensiones del base de datos
tabla.add_row(['DIMENSIONES', f'{filas} filas', f'{columnas} columnas'])

# Mostrar la tabla
print(tabla)

### Licencias de uso de datos

Los datos utilizados en este análisis provienen de los siguientes recursos:

- **Estadística General de Incendios Forestales (EGIF)**, gestionada por el Ministerio para la Transición Ecológica y el Reto Demográfico (MITECO). Los datos deben ser utilizados de acuerdo con las [condiciones de uso de la EGIF](https://www.miteco.gob.es/es/biodiversidad/temas/incendios-forestales/estadisticas-datos.aspx).
- **Nomenclátor Geográfico de Municipios y Entidades de Población (NGMEP)**, proporcionado por el **Centro Nacional de Información Geográfica (CNIG)**. Los datos están sujetos a la licencia de [Datos Abiertos del CNIG](https://astronoomia.ign.es/web/ign/portal). Se debe proporcionar citación adecuada al utilizar los datos.
- **Código INE** proporcionado por el **Instituto Nacional de Estadística (INE)**. Los datos están sujetos a las [condiciones de uso del INE](https://www.ine.es/jaxiT3/Datos.htm?t=2865), y deben ser utilizados con la debida citación.

Recuerda que es importante respetar las licencias de uso al compartir o distribuir los resultados de este notebook.
