==> 02_Preprocesamiento_extraccion_datos_climatologicos_AEMET.ipynb <==
import pandas as pd
import numpy as np
import json
from datetime import datetime
import os
from geopy.distance import geodesic
import requests
import time

## Extracción de datos históricos de los incendios por fecha de AEMET

Cargamos la última versión del dataset

bdif = pd.read_excel("bdif_geograficas.xlsx")
bdif.head()

bdif.shape

Extraemos las fechas de los incendios

# Guardamos las fechas de detección de los incendios
fechas = pd.to_datetime(bdif['deteccion'])
# Extraemos solo la fecha
fechas = fechas.dt.date
# Eliminamos los duplicados 
fechas = fechas.drop_duplicates()
fechas.head()

Definimos la clave API de AEMET y configuramos la ruta de alamacenamiento

# API key de AEMET
api_key = ''
# Ruta a la carpeta de almacenamiento
directorio = 'meteo/'

Extraemos los datos

# Función para extraer los datos
def extraer_meteo(fecha):
    """
    Función para extraer los datos meteorológicos diarios de AEMET para una fecha específica.
    
    Esta función realiza una solicitud a la API de AEMET para obtener datos climatológicos
    de una fecha de todas las estaciones meteorológicas.
    Los datos se descargan en formato JSON y se guardan con la fecha como nombre.
    
    Parámetros:
    fecha (str): Fecha del incendio en formato 'YYYY-MM-DD'.

    La función hace lo siguiente:
    1. Construye la fecha de inicio y fin con formato específico requerido por la API de AEMET.
    2. Realiza una solicitud a la API de AEMET con las fechas de inicio y fin para obtener los enlaces de los datos.
       Como las fechas de inicio y fin son iguales, se ajusta la hora para realizar la consulta correctamente.
    3. Descarga los datos en formato JSON.
    4. Guarda los datos descargados en un archivo con el nombre correspondiente a la fecha.
    
    Retorna:
    Se van guardando los archivos y se obtienen mensajes por consola.
    """
    # Montamos la fecha de petición en el formato que se solicita
    fechaini = str(fecha) + 'T00:00:00UTC'
    fechafin = str(fecha) + 'T23:59:59UTC'
    
    # Realizamos la solicitud
    url = f'https://opendata.aemet.es/opendata/api/valores/climatologicos/diarios/datos/fechaini/{fechaini}/fechafin/{fechafin}/todasestaciones'
    headers = {'cache-control': "no-cache"}
    response = requests.request("GET", url, headers=headers, params={"api_key":api_key})

    # Comprobamos que la respuesta sea ok
    if response.status_code == 200:
        # Obtener el enlace de los datos
        data_url = response.json().get('datos', '')

        # Descargar el json
        if data_url:
            data_response = requests.get(data_url)
            if data_response.status_code == 200:
                # Guardar el json con la fecha como nombre
                file_path = f"{directorio}/{str(fecha)}.json"
                with open(file_path, 'w') as f:
                    f.write(data_response.text)
                print(f'Datos descargados para la fecha {str(fecha)}')
            else:
                print(f'Error al descargar los datos para la fecha {str(fecha)}')
        else:
            print(f'No se encontraron datos para la fecha {str(fecha)}')
    else:
        print(f'Error en la solicitud para la fecha {str(fecha)}: {response.status_code}')

# Limitamos la conexión porque solo podemos hacer 50 peticiones por minuto
for fecha in fechas:
    extraer_meteo(fecha)
    time.sleep(2)

## Listado de estaciones meteorológicas

# Cargamos el archivo
estaciones = pd.read_json('estaciones_meteorologicas.json')

# Mostramos los primeros registros
estaciones.head()

# Vamos a extraer el indicativo y las coordenadas de cada estación meteorológica
estaciones = estaciones[['indicativo', 'latitud', 'longitud']]
estaciones.head()

# Necesitamos las coordenadas en grados
# Creamos una función de transformación
def conversion_a_grados(coordenada):
    """
    Dada una coordenada en grados, minutos y segundos, la transformamos a grados decimales.

    La entrada es una cadena de texto que tiene la siguiente estructura:
    - Los dos primeros numeros que representan los grados.
    - Los dos números centrales que representan los minutos.
    - Los dos últimos números representan los segundos.
    - La letra final que representa la orientación: 'N', 'S', 'E' o 'W'.
    
    Ejemplo de entrada: '394924N'

    La fórmula que realiza la conversión es la siguiente:
    grados + (minutos / 60) + (segundos / 3600).

    Si tenemos orientación 'S' o 'W' el valor será negativo.

    Parámetros:
    coordenada (str): coordenada en grados, minutos y segundos más la orientación.

    Retorna:
    float: La coordenada en grados decimales.

    Ejemplo:
    conversion_a_grados('394924N') -> 39.8233333
    conversion_a_grados('074550W') -> -74.9166667
    """
    g = int(coordenada[:2])
    m = int(coordenada[2:4])
    s = int(coordenada[4:6])  
    o = coordenada[6]
    gd = g + (m/60) + (s/3600)
    if o in ['S', 'W']:
        gd = -gd
    return gd

# Le pasamos la función a las columnas de latitud y longitud
estaciones['latitud'] = estaciones['latitud'].apply(conversion_a_grados)
estaciones['longitud'] = estaciones['longitud'].apply(conversion_a_grados)

# Mostramos el dataframe para comprobar los cambios
estaciones.head()

## Búsqueda de la estación más cercana

# Creamos la función que dada unas coordenadas encuentre las estaciones
# meteorológica más cercanas por orden de proximidad
def estacion_cercana(latitud, longitud):
    """
    Encuentra las estaciones meteorológicas más cercanas dadas unas coordenadas geográficas.

    Calcula la distancia entre las coordenadas dadas y las de cada estación y devuelve
    un dataframe con las estaciones ordenadas por distancia ascendente.

    Parámetros:
        latitud (float): Latitud de la ubicación del incendio.
        longitud (float): Longitud de la ubicación del incendio.

    Retorna:
        pandas.DataFrame: DataFrame con las estaciones más cercanas. Incluye:
            - 'indicativo': un código que identifica la estación.
            - 'distancia': la distancia en km de las coordenadas a la estación.
    """
    estaciones_cercanas = []
    for i, f in estaciones.iterrows():
        # Calculo de la distancia entre las coordenadas iniciales y la estación correspondiente
        distancia = geodesic( (latitud,longitud),(f['latitud'],f['longitud'])).kilometers

        # Vamos guardando las estaciones y las distancias
        estaciones_cercanas.append({
            'indicativo': f['indicativo'],
            'distancia': distancia
        })

    # Lo pasamos a dataframe
    df_estaciones_cercanas = pd.DataFrame(estaciones_cercanas)

    # Lo ordenamos por distancia ascendente
    df_estaciones_cercanas = df_estaciones_cercanas.sort_values(by='distancia', ascending=True).reset_index(drop=True)

    return df_estaciones_cercanas

## Extracción de los datos meteorológicos

# Creamos una función que dada una fecha y unas coordenadas,
# extraiga los datos meteorológicos de la carpeta donde se han guardado
# el historico de datos climatologicos de las fechas de los incendios

def extraer_datos (fecha, indicativo):
    """
    Extrae los datos meteorológicos de una estación específica de un archivo JSON para una fecha dada.

    Args:
        fecha (str): Fecha del archivo JSON a leer (formato YYYY-MM-DD).
        indicativo (str): Indicativo de la estación meteorológica.

    Returns:
        dict: Datos meteorológicos de la estación si se encuentra, o None si no se encuentra.
    """
    # Ruta al archivo
    archivo = os.path.join('meteo', f'{fecha}.json')
    # Leemos el json y cargamos los datos
    datos = pd.read_json(archivo,  encoding='latin-1')
    
    # Buscamos si la estación está en el archivo
    for _,estacion in datos.iterrows():
        if estacion['indicativo'] == indicativo:
            
            # Convertir la fila de la estación en un diccionario
            estacion = estacion.to_dict()  # Convertir la fila en un diccionario
            
            # Convertir las claves del diccionario a minúsculas
            estacion = {k.lower(): v for k, v in estacion.items()}
            
            # Extraer los datos meteorológicos de la estación
            meteo = {"tmed": estacion.get("tmed", ""),
                    "prec": estacion.get("prec", ""),
                    "tmin": estacion.get("tmin", ""),
                    "tmax": estacion.get("tmax", ""),
                    "dir": estacion.get("dir", ""),
                    "velmedia": estacion.get("velmedia", ""),
                    "racha": estacion.get("racha", ""),
                    "sol": estacion.get("sol", ""),
                    "presmax": estacion.get("presmax", ""),
                    "presmin": estacion.get("presmin", ""),
                    "hrmedia": estacion.get("hrmedia", ""),
                    "hrmax": estacion.get("hrmax", ""),
                    "hrmin": estacion.get("hrmin", "")}
            return meteo
    return None

# Ahora cargamos el archivo con el numero de parte, fecha y coordenadas
partes = bdif[['parte', 'deteccion', 'lon', 'lat']].sort_values(by='deteccion', ascending=True)
partes.head()

## Selección de fechas

Elegimos el intervalo de fechas a descargar, por ejemplo, de año en año

# Filtro de fechas
inicio = pd.to_datetime('2021-01-01')
fin = pd.to_datetime('2021-12-31')

# Convertir la columna 'fecha' a formato datetime
partes['deteccion'] = pd.to_datetime(partes['deteccion'])

# Filtrar por rango
filtro = (partes['deteccion'] >= inicio) & (partes['deteccion'] <= fin)
partes = partes[filtro]
partes.head()

clima = []
start_time = time.time()  # Guardamos el tiempo inicial

# Vamos recorriendo los incendios
for _, parte in partes.iterrows():
    # Creamos el registro del incendio
    registro_clima = {'parte': parte['parte'], 'indicativo': indicativo}
    
    # Tomamos la fecha, longitud y latitud
    fecha = pd.to_datetime(parte['deteccion']).strftime('%Y-%m-%d')
    lon = parte['lon']
    lat = parte['lat']
    
    # Obtenemos el listado de estaciones más cercanas al incendio
    lista_estaciones_cercanas = estacion_cercana(lat, lon)

    # Extraemos la información
    for _, estacion in lista_estaciones_cercanas.iterrows():
        indicativo = estacion['indicativo']
        meteo = extraer_datos(fecha, indicativo)
        # En el momento que tiene la información sale
        if meteo:
            break

    # Agregamos los datos al registro creado
    registro_clima.update(meteo)

    # Añadir el registro a la lista
    clima.append(registro_clima)


# Convertir la lista en dataframe
bdif_climatologicas = pd.DataFrame(clima)
bdif_climatologicas

end_time = time.time()  # Guardamos el tiempo final

execution_time = end_time - start_time  # Calculamos el tiempo de ejecución
print(f"Tiempo de ejecución: {execution_time} segundos")

# Extraemos los datos a excel
bdif_climatologicas.to_excel("variables climatologicas/bdif_climatologicas_2021.xlsx", index=False)

## Fusión de archivos

# Listado de años
años = range(1974, 2021)

dataframes = []

# Vamos cargando los archivos
for año in años:
    # Montamos la ruta al archivo
    archivo = f'variables climatologicas/bdif_climatologicas_{año}.xlsx'
    
    df_año = pd.read_excel(archivo)
        
    # Añadir el DataFrame cargado a la lista
    dataframes.append(df_año)

# Concatenamos todos
bdif_climatologicas = pd.concat(dataframes, ignore_index=True)

# Mostrar el DataFrame concatenado
bdif_climatologicas.head()

bdif_climatologicas.to_excel('bdif_climatologicas.xlsx', index=False)
print(f"El archivo ha sido guardado como {archivo_salida}")

# Cargar el dataset y mostrar las primeras filas
bdif_geograficas = pd.read_excel("bdif_geograficas.xlsx")
bdif_geograficas.head()

# Cargar el dataset y mostrar las primeras filas
bdif_climatologicas = pd.read_excel("bdif_climatologicas.xlsx")
bdif_climatologicas.head()

# Unimos los dataframes
bdif = pd.merge( bdif_geograficas, bdif_climatologicas, on='parte', how='left')

# Mostramos las primeras filas como comprobación
bdif.head()

bdif.to_excel('bdif_geo_clim.xlsx', index=False)
print(f"El archivo ha sido guardado como bdif_geo_clim")

bdif.shape

bdif.info()

### Licencias de uso de datos

Los datos utilizados en este análisis provienen de los siguientes recursos:

- **Estadística General de Incendios Forestales (EGIF)**, gestionada por el Ministerio para la Transición Ecológica y el Reto Demográfico (MITECO). Los datos deben ser utilizados de acuerdo con las [condiciones de uso de la EGIF](https://www.miteco.gob.es/es/biodiversidad/temas/incendios-forestales/estadisticas-datos.aspx).
- **AEMET** y están sujetos a las [condiciones de uso de AEMET](https://www.aemet.es/es/serviciosclimaticos/datos).
- **Nomenclátor Geográfico de Municipios y Entidades de Población (NGMEP)**, proporcionado por el **Centro Nacional de Información Geográfica (CNIG)**. Los datos están sujetos a la licencia de [Datos Abiertos del CNIG](https://astronoomia.ign.es/web/ign/portal). Se debe proporcionar citación adecuada al utilizar los datos.

Recuerda que es importante respetar las licencias de uso al compartir o distribuir los resultados de este notebook.
