==> 03_Preprocesamiento_extraccion_datos_FWI.ipynb <==
import xarray as xr
import cartopy.crs as ccrs
import numpy as np
import time
import pandas as pd

pip install netCDF4

# Ruta al archivo .nc
file_path = "fwi/1974.nc"  # Cambia esto al nombre del archivo

# Cargar el archivo NetCDF
dataset = xr.open_dataset(file_path)

# Ver información general del archivo
print(dataset)

def obtener_fwi(dataset, fecha, lat_query, lon_query):
    """
    Obtiene el valor del Fire Weather Index (FWI) para una fecha y coordenadas específicas del dataset.
    
    Parámetros:
    dataset: xarray.Datasetm, contiene el FWI y coordenadas
    fecha: str o np.datetime64, fecha en formato 'YYYY-MM-DD'
    lat_query: float, latitud
    lon_query: float, longitud

    Retorna: float, el valor de FWI en las coordenadas especificadas para la fecha dada
    """
    
    # Seleccionar los datos del FWI para la fecha dada
    fwi_dia = dataset['fwi-daily-proj'].sel(time=fecha)
    
    # Obtenemos latitudes y longitudes
    latitudes = fwi_dia['lat'].values
    longitudes = fwi_dia['lon'].values

    # Inicializamos una variable que guarda la minima distancia 
    min_distance = float('inf')
    lat_index, lon_index = None, None

    # Recorremos las cuadriculas
    for i in range(latitudes.shape[0]):
        for j in range(longitudes.shape[1]):
            distance = np.sqrt((latitudes[i, j] - lat_query) ** 2 + (longitudes[i, j] - lon_query) ** 2)
            
            if distance < min_distance:
                min_distance = distance
                lat_index, lon_index = i, j

    # Si lo sindices son validos
    if lat_index is not None and lon_index is not None:
        fwi_value = fwi_dia.values[lat_index, lon_index]
        
        lat_value = latitudes[lat_index, lon_index]
        lon_value = longitudes[lat_index, lon_index]
        
        return fwi_value, lat_value, lon_value
    else:
        raise ValueError("No se encontraron valores")

bdif = pd.read_excel("bdif_geograficas.xlsx")
bdif.head()

# Ahora cargamos el archivo con el numero de parte, fecha y coordenadas
partes = bdif[['parte', 'deteccion', 'lon', 'lat']].sort_values(by='deteccion', ascending=True)
partes.head()

# Filtro de fechas
inicio = pd.to_datetime('1988-01-01')
fin = pd.to_datetime('1988-12-31')

# Convertir la columna 'fecha' a formato datetime
partes['deteccion'] = pd.to_datetime(partes['deteccion'])

# Filtrar por rango
filtro = (partes['deteccion'] >= inicio) & (partes['deteccion'] <= fin)
partes = partes[filtro]
partes.head()

fwi = []
start_time = time.time()

# Cargar el dataset
dataset = xr.open_dataset('fwi/1988.nc')

# Recorremos los incendios para obtener el FWI para cada uno
for _, parte in partes.iterrows():
    # Creamos el registro del incendio
    registro_fwi = {'parte': parte['parte']}
    
    # Cogemos la fecha, longitud y latitud
    fecha = pd.to_datetime(parte['deteccion']).strftime('%Y-%m-%d')
    lon = parte['lon']
    lat = parte['lat']

    # Obtener el valor del FWI
    try:
        fwi_value, lat_value, lon_value = obtener_fwi(dataset, fecha, lat, lon)
        registro_fwi['fwi'] = fwi_value
    
    except ValueError as e:
        print(f"Error al obtener FWI para el incendio {parte['parte']}: {e}")
        continue

    fwi.append(registro_fwi)

bdif_fwi = pd.DataFrame(fwi)

bdif_fwi.head()
bdif_fwi.to_excel("fwi/bdif_fwi_1988.xlsx", index=False)

end_time = time.time()
execution_time = (end_time - start_time)/60
print(f"Tiempo de ejecución: {execution_time} minutos")

### Fusión de archivos

# Listado de años
años = range(1974, 2021)

dataframes = []

# Vamos cargando los archivos
for año in años:
    # Montamos la ruta al archivo
    archivo = f'fwi/bdif_fwi_{año}.xlsx'
    
    df_año = pd.read_excel(archivo)
        
    dataframes.append(df_año)

bdif_fwi = pd.concat(dataframes, ignore_index=True)

# Mostrar el df
bdif_fwi.head()

bdif_fwi.to_excel('bdif_fwi.xlsx', index=False)
print(f"El archivo ha sido guardado como bdif_fwi")

# Cargar el dataset y mostrar las primeras filas
bdif_geo_clim = pd.read_excel('bdif_geo_clim.xlsx')
bdif_geo_clim.head()

# Unimos los dataframes
bdif_geo_clim_fwi = pd.merge(bdif_geo_clim, bdif_fwi, on='parte', how='left')

# Mostramos las primeras filas como comprobación
bdif_geo_clim_fwi.head()

bdif_geo_clim_fwi.to_excel('bdif_geo_clim_fwi.xlsx', index=False)
print(f"El archivo ha sido guardado como bdif_geo_clim_fwi")

### Licencias de uso de datos

Los datos utilizados en este análisis provienen de los siguientes recursos:

- **Estadística General de Incendios Forestales (EGIF)**, gestionada por el Ministerio para la Transición Ecológica y el Reto Demográfico (MITECO). Los datos deben ser utilizados de acuerdo con las [condiciones de uso de la EGIF](https://www.miteco.gob.es/es/biodiversidad/temas/incendios-forestales/estadisticas-datos.aspx).
- **Climate Data Store (CDS)**, específicamente del indicador de peligro de incendio **FWI** derivado de proyecciones climáticas para Europa. Los datos están sujetos a la [licencia de uso de CDS](https://cds.climate.copernicus.eu/#!/home), y deben ser utilizados con la debida citación.
- **Nomenclátor Geográfico de Municipios y Entidades de Población (NGMEP)**, proporcionado por el **Centro Nacional de Información Geográfica (CNIG)**. Los datos están sujetos a la licencia de [Datos Abiertos del CNIG](https://astronoomia.ign.es/web/ign/portal). Se debe proporcionar citación adecuada al utilizar los datos.

Recuerda que es importante respetar las licencias de uso al compartir o distribuir los resultados de este notebook.
