==> 04_Preprocesamiento_estimacion_poblacion.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from scipy.optimize import curve_fit

### **Estimación de la población de los municipios a lo largo de los años**

Dado que en el dataset de EGIF hemos introducido el dato de la población actual de los municipios y ha resultado ser relevante para el modelado, hacemos una revisión trando de introducir una estimación más realista de la población en función de los años, puesto que la población ha variado inevitablemente desde 1974.

Para ello se ha ajustado una ecuación lineal a la evolución de la población para cada una de las provincias, para después mediante proporción extrapolarlo a la población de cada uno de los municpios de la provincia correspondiente. A continuación, se detalla el proceso que se ha seguido.

### **Uso del modelo de ecuación lineal de la provincia**

El modelo provincial permite estimar la población de una provincia en cualquier año utilizando la fórmula de una ecuación lineal:

$$
\text{Población Estimada Provincia} = m \cdot (\text{Año}) + b
$$

Donde:
- $m$: Pendiente de la recta que refleja la tasa de cambio anual de la población.
- $b$: Intersección con el eje $y$
- $Año$: Año para el que queremos calcular la población.

### **Estimación de la población del municipio mediante proporcionalidad**
Para estimar la población de un municipio utilizamos los datos del último año de población conocidos y los ajustamos proporcionalmente a la evolución que ha sufrido la población de la provincia suponiendo que la relación entre el municipio y la provincia sea constante a lo largo del tiempo. La fórmula utilizada es la siguiente:

$$
\text{Población estimada municipio} = \text{Población municipio} \cdot \frac{\text{Población estimada provincia}}{\text{Población provincia}}
$$

Donde:
- **Población municipio**: Población actual del municipio
- **Población estimada provincia**: Estimación de la población de la provincia para el año seleccionado.
- **Población provincia**: Problación actual de la provincia.


# Cargar los datos de evolución de la población
poblacion = pd.read_excel('evolucion_poblacion.xlsx')
poblacion.head()

# Cargamos los datos del dataset del egif
bdif = pd.read_excel('bdif_geo_clim_fwi.xlsx')
bdif.head()

# Hay que trasponer los años y la población 
poblacion = poblacion.melt(id_vars=['cod_prov', 'Provincias'], var_name='Año', value_name='Población')

# Pasamos año a númerico
poblacion['Año'] = poblacion['Año'].astype(int)

# Mostramos
poblacion.head()

# Definimos la ecuación lineal con la que modelamos la evolución de la población
def ecuacion_lineal(x, m, b):
    return m * x + b

# Creamos un diccionario vacío que guarde la ecuación de evolución de la población para cada provincia
ecuacion_provincia = {}

for cod in poblacion['cod_prov'].unique():
    # Recuperamos los datos de población para cada año de la provincia
    datos = poblacion[poblacion['cod_prov'] == cod]
    fechas = datos['Año'].values
    poblaciones = datos['Población'].values

    # Obtenemos la ecuación lineal y guardarmos los parámetros, obviamos la varianza
    parametros, _ = curve_fit(ecuacion_lineal, fechas, poblaciones)
    ecuacion_provincia[cod] = parametros

# Mostramos el resulado
print(ecuacion_provincia)

# Población base de las provincias la obtenemos del dataset de población para el último año disponible
poblacion_base_provincia = poblacion[poblacion['Año'] == 2022].set_index('cod_prov')['Población'].to_dict()

# Creamos una función que haga una estimación de la población del municipio en función del que se hace en la provincia
def estimar_poblacion(fila, ecuacion_provincia, poblacion_provincia):
    """
    Calcula la población estimada de un municipio basado en su proporción con la población de la provincia y la estimación
    de la población de la provincia.

    Parametros:
    fila: fila del dataframe de donde extraemos el año, el codigo de la provincia y la poblacion del municipio
    ecuacion_provincia: diccionario que contiene las ecuaciones lineales de evolución de la población por provincia
    poblacion_provincia: diccionario que contiene las poblaciones actuales de las provincias
    
    Retorna:
    float: Población estimada para el municipio.
    """
    cod_prov = fila['cod_prov']
    ano = fila['año']
    poblacion_base_mun = fila['poblacion']
    
    if cod_prov in ecuacion_provincia:
        # Obtenemos la ecuación lineal de la provincia
        m, b = ecuacion_provincia[cod_prov]
        
        # Calculamos la población estimada en el año correspondiente
        poblacion_est_prov = m * ano + b
        
        # Obtener población base de la provincia
        poblacion_base_prov = poblacion_provincia.get(cod_prov)
        
        # Calcular población estimada del municipio por proporcionalidad con la de la provincia
        poblacion_estimada = poblacion_base_mun * (poblacion_est_prov / poblacion_base_prov)
        
        # Si la población estimada es negativa, asignar 1
        if poblacion_estimada < 0:
            poblacion_estimada = 1
        
        return poblacion_estimada

# Aplicar la función al dataset bdif
bdif['poblacion_estimada'] = bdif.apply(estimar_poblacion, axis=1, args=(ecuacion_provincia, poblacion_base_provincia))

# Mostramos los resultados
bdif.head()

bdif['poblacion_estimada'].describe()

bdif.to_excel("bdif_geo_clim_fwi_pob.xlsx", index=False)

### Licencias de uso de datos

Los datos utilizados en este análisis provienen de los siguientes recursos:

- **Estadística General de Incendios Forestales (EGIF)**, gestionada por el Ministerio para la Transición Ecológica y el Reto Demográfico (MITECO). Los datos deben ser utilizados de acuerdo con las [condiciones de uso de la EGIF](https://www.miteco.gob.es/es/biodiversidad/temas/incendios-forestales/estadisticas-datos.aspx).
- **Series de población por provincia (1900-2001)** y las **principales series de población desde 1998** proporcionadas por el **Instituto Nacional de Estadística (INE)**. Los datos están sujetos a las [condiciones de uso del INE](https://www.ine.es/jaxiT3/Datos.htm?t=2865), y deben ser utilizados con la debida citación.
- **Nomenclátor Geográfico de Municipios y Entidades de Población (NGMEP)**, proporcionado por el **Centro Nacional de Información Geográfica (CNIG)**. Los datos están sujetos a la licencia de [Datos Abiertos del CNIG](https://astronoomia.ign.es/web/ign/portal). Se debe proporcionar citación adecuada al utilizar los datos.

Recuerda que es importante respetar las licencias de uso al compartir o distribuir los resultados de este notebook.
